Move the innerHTML patch to a separate method to be used by both Element and ShadowRoot.
